refactor rename properties
